+++ src/Components/Input.js
@@ -10,6 +10,7 @@ const Input = props => {
         name,
         label,
         action,
+        actionLocation,
         helpText,
         error,
         state,
@@ -54,26 +55,41 @@ const Input = props => {
         (state === 'error') ? 'border-red-500 text-red-500 focus:border-red-500 dark:focus:border-red-500 focus:ring-red-500 dark:focus:ring-red-500' : '',
         (state === 'saved') ? 'border-green-500 text-green-500 focus:border-green-500 dark:focus:border-green-500 focus:ring-green-500 dark:focus:ring-green-500' : '',
         (state === 'saving') ? 'border-orange-500 focus:border-orange-500 dark:focus:border-orange-500 focus:ring-orange-500 dark:focus:ring-orange-500' : '',
+        (actionLocation === 'left') ? 'pl-10' : '',
         inputClassName,
     )
 
     const Append = () => {
         if (state === 'saving') {
-            return <Loader style="circle" height="5" width="5" color="orange" />
+            return (
+                <div className={`flex absolute inset-y-0 ${actionLocation}-0 items-center p${actionLocation.charAt(0)}-3`}>
+                    <Loader style="circle" height="5" width="5" color="orange" />
+                </div>
+            )
         } else if (state === 'saved') {
             return (
-                <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-green-500" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" />
-                </svg>
+                <div className={`flex absolute inset-y-0 ${actionLocation}-0 items-center p${actionLocation.charAt(0)}-3`}>
+                    <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-green-500" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" />
+                    </svg>
+                </div>
             )
         } else if (state === 'error') {
             return (
-                <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-red-500" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                </svg>
+                <div className={`flex absolute inset-y-0 ${actionLocation}-0 items-center p${actionLocation.charAt(0)}-3`}>
+                    <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-red-500" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
+                    </svg>
+                </div>
             )
-        } else if (action) {
+        } else if (action && actionLocation === 'none') {
             return action;
+        } else if (action) {
+            return (
+                <div className={`flex absolute inset-y-0 ${actionLocation}-0 items-center p${actionLocation.charAt(0)}-3`}>
+                    {action}
+                </div>
+            )
         }
         return null;
     }
@@ -85,9 +101,7 @@ const Input = props => {
             }
             <div className="relative">
                 <input id={id} name={name} className={inputClasses} {...attributes} />
-                <div className="flex absolute inset-y-0 right-0 items-center pr-3">
-                    <Append />
-                </div>
+                <Append />
             </div>
             {
                 (helpText !== '' || error !== '')
@@ -106,6 +120,7 @@ Input.propTypes = {
     name: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
     label: PropTypes.string,
     action: PropTypes.object,
+    actionLocation: PropTypes.oneOfType(['right', 'left', 'none']),
     helpText: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
     error: PropTypes.string,
     state: PropTypes.oneOf(['', 'saving', 'error', 'saved']),
@@ -118,6 +133,7 @@ Input.defaultProps = {
     error: '',
     state: '',
     label: '',
+    actionLocation: 'right',
     helpText: '',
     wrapperClassName: 'mb-6',
     labelClassName: '',